refactor: move namespace message to inside of namespace channel, move discord-public-message.cpp contents to discord-public-channel.cpp
